% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{AF16xLSJ2markers}
\alias{AF16xLSJ2markers}
\title{Marker/position lookup table for AF16xLSJ2 RIAILs}
\format{A data frame with 1031 rows and 4 variables:
\describe{
  \item{marker}{the name of the marker used for genotyping}
  \item{chr.num}{the chromosome number in arabic numerals}
  \item{chr.roman}{the chromosome number in roman numerals}
  \item{chr.roman}{the position of the marker in genome build cb4}
}}
\description{
This data frame contains a lookup table for the translation of marker
positions from genetic to physical positions. The physical positions in this
table come from C. briggsae genome build cb4.
}

