% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{AllCBfosmids}
\alias{AllCBfosmids}
\title{Dataframe of all CB4856 fosmids in the Andersen Lab}
\format{A data frame with 11453 rows and 5 variables:
\describe{
  \item{chr}{the chromosome (in roman numerals) on which the fosmid is found}
  \item{clone}{the name of the fosmid}
  \item{start}{the genetic position at which the fosmid begins}
  \item{end}{the genetic position at which the fosmid ends}
  \item{newpos}{the current location of the fosmid in the Andersen Lab}
}}
\description{
This data frame contains genomic position and laboratory stock
information for all CB4856 fosmids.
}

