% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{annotate_lods}
\alias{annotate_lods}
\title{Convert jb LODmatrix to scanone object}
\usage{
annotate_lods(lods, cross)
}
\arguments{
\item{cross}{An example cross object from which to extract scanone skeleton}

\item{LODS}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{LL}{}
}
\value{
The genotype matrix, encoded as -1 or 1 for genotype
}
\description{
Convert jb LODmatrix to scanone object
}

