% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{calc_mediation}
\alias{calc_mediation}
\title{Calculate mediation scores between phenotype and gene expression}
\usage{
calc_mediation(peak, expression_probe, phenodf, cross, scaled = T, lm = FALSE)
}
\arguments{
\item{peak}{peak marker of phenotype QTL in format chr:pos}

\item{phenodf}{Dataframe for phenotype in the form of strain, trait, phenotype with the strain matching RIL from set1 (where we have expression data)}

\item{cross}{Empty cross object, used to gather genotype information (does not need to be merged with phenotype, but can be)}

\item{scaled}{Boolean noting if the phenotype should be scaled for a mean of 0 and standard devitaiton of 1. Default is TRUE}

\item{lm}{Boolean argument. If TRUE, a set of linear models will be used to calculate mediation. If FALSE, mediation will be calculated from the R package "mediation"}

\item{probe}{probeID from expression data}
}
\description{
Calculate mediation scores between phenotype and gene expression
}
