% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{cint}
\alias{cint}
\title{Calculate the variance explained, effect size, and confidence interval bounds
for each peak}
\usage{
cint(lods, chr, lodcolumn = 5, drop = 1.5)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{chr}{An example cross object from which to extract scanone skeleton}

\item{lodcolumn}{The index of the column containing the lods scores}

\item{drop}{The LOD drop for calculating the confidence interval}
}
\value{
The annotated LOD data frame with added columns for peak information
}
\description{
Calculate the variance explained, effect size, and confidence interval bounds
for each peak
}

