% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{corplot}
\alias{corplot}
\title{Test for correlation among traits in a data frame and output a plot}
\usage{
corplot(df, groups, phenocol = NA)
}
\arguments{
\item{df}{A dataframe output by easysorter}

\item{phenocol}{When spread, the column number on which traits begin, usually 11 with HTA data.}

\item{gropus}{Number of groups to split the correlation into, usually based on number of significant PCs}
}
\value{
A dendrogram and a correlation plot for the data
}
\description{
Test for correlation among traits in a data frame and output a plot
}
