% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{eqtlpheno}
\alias{eqtlpheno}
\title{Expression phenotypes from Rockman 2010 paper, reanalyzed for Evans and Andersen 2020}
\format{A data frame with 4 columns and 2948363 rows}
\description{
Expression phenotypes from Rockman 2010 paper, reanalyzed for Evans and Andersen 2020
}
