% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LOD.R
\name{fsearch}
\alias{fsearch}
\title{Map all of the traits in a given cross object with forward search}
\usage{
fsearch(cross, phenotype = NULL, iterations = 1000, doGPU = FALSE)
}
\arguments{
\item{cross}{A complete cross object with the phenotype data merged}

\item{iterations}{The number of iterations for the FDR calculation. Defaults
to \code{1000}.}

\item{doGPU}{Boolean, whether to use the gputools package to speed up,
mapping. This can only be set to \code{TRUE} on machines with an NVIDEA
graphics card with the gputools package installed. Defaults to \code{FALSE}.}
}
\value{
The LOD scores for all markers
}
\description{
Map all of the traits in a given cross object with forward search
}

