% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{get_chr_peaks}
\alias{get_chr_peaks}
\title{Convert LODmatrix to scanone object}
\usage{
get_chr_peaks(mindex.split, chr.mindex.offset, LODS)
}
\arguments{
\item{LODS}{A data frame output by the mapping functions to be converted to a}

\item{cross}{An example cross object from which to extract scanone skeleton}
}
\value{
The genotype matrix, encoded as -1 or 1 for genotype
}
\description{
Convert LODmatrix to scanone object
}

