% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LOD.R
\name{get_lod_by_cor}
\alias{get_lod_by_cor}
\title{Get the LOD value for each marker based on the correlation between genotype
and phenotype at each marker.}
\usage{
get_lod_by_cor(npheno, pheno, gdata, doGPU = FALSE)
}
\arguments{
\item{npheno}{}

\item{pheno}{}

\item{doGPU}{Boolean, whether to use the gputools package to speed up,
mapping. This can only be set to \code{true} on machines with an NVIDEA
graphics card with the gputools package installed.}

\item{geno}{}
}
\value{
The LOD scores for all markers
}
\description{
Get the LOD value for each marker based on the correlation between genotype
and phenotype at each marker.
}

