% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{get_peak_fdr}
\alias{get_peak_fdr}
\title{Convert jb LODmatrix to scanone object}
\usage{
get_peak_fdr(chromosome.peaks.lod, pdata, gdata, perms = 100, doGPU = F)
}
\arguments{
\item{LODS}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{cross}{An example cross object from which to extract scanone skeleton}

\item{LL}{}
}
\value{
The genotype matrix, encoded as -1 or 1 for genotype
}
\description{
Convert jb LODmatrix to scanone object
}

