% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{indels}
\alias{indels}
\title{Dataframe of indels between N2 and CB4856}
\format{A data frame with 10257 rows and 15 variables}
\description{
This data frame contains position information for all indels between
N2 and CB4856 greater than 25 bp. Developed by the Kammenga lab.
}

