% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{load_cross_obj}
\alias{load_cross_obj}
\title{Load Cross Obj}
\usage{
load_cross_obj(name)
}
\arguments{
\item{name}{Name of cross object}
}
\value{
A cross object
}
\description{
Download large cross objects if need be and load or load local cross obj.
}

