% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{lodplot}
\alias{lodplot}
\title{Plot all lod scores for a forward search mapping by iteration}
\usage{
lodplot(map)
}
\arguments{
\item{map}{The mapping output for a single trait}
}
\value{
A plot with a line for each iteration of the forward search mapping
}
\description{
Plot all lod scores for a forward search mapping by iteration
}

