% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LOD.R
\name{map}
\alias{map}
\title{Map all of the traits in a given cross object}
\usage{
map(cross, doGPU = FALSE)
}
\arguments{
\item{cross}{A complete cross object with the phenotype data merged}

\item{doGPU}{Boolean, whether to use the gputools package to speed up,
mapping. This can only be set to \code{true} on machines with an NVIDEA
graphics card with the gputools package installed.}
}
\value{
The LOD scores for all markers
}
\description{
Map all of the traits in a given cross object
}

