% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{maxpeaks}
\alias{maxpeaks}
\title{Get max LOD score and SNP index for each mapped trait}
\usage{
maxpeaks(lods, cross)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a}

\item{cross}{An example cross object from which to extract scanone skeleton}
}
\value{
A list consisting of two vectors, the first being the max peak LOD
height and the second being the index of the SNP at which the peak LOD occurs
}
\description{
Get max LOD score and SNP index for each mapped trait
}

