% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{mergepheno}
\alias{mergepheno}
\title{Merge the cross object and the phenotype data frame using a dplyr left_join}
\usage{
mergepheno(cross, phenotype, set = NULL)
}
\arguments{
\item{cross}{A cross object}

\item{phenotype}{The phenotype data frame with the id numbers for each strain}

\item{set}{Filter the phenotype data to one specific set (Rockman=1 or 2,
Andersen=3) before joining to the data frame}
}
\value{
The phenotype element of the cross object
}
\description{
Merge the cross object and the phenotype data frame using a dplyr left_join
}

