% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pxgplot}
\alias{pxgplot}
\title{Plot the phenotype by genotype split at each peak marker in a mapping}
\usage{
pxgplot(cross, map, parent = "N2xCB4856", tsize = 20)
}
\arguments{
\item{cross}{The cross object used for the mapping}

\item{map}{The mapping output for a single trait}

\item{parent}{The parental strains used to generate RIAILs. Either "N2xCB4856"}

\item{tsize}{The theme size of the text for the plot, default is 20.
(default), "N2xLSJ2", or "AF16xHK104"}
}
\value{
A boxplot of the phenotype by genotype split at each peak marker in a
mapping
}
\description{
Plot the phenotype by genotype split at each peak marker in a mapping
}
