% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pxgplot_par}
\alias{pxgplot_par}
\title{Plot the phenotype by genotype split at each peak marker in a mapping
next to parental phenotype split}
\usage{
pxgplot_par(cross, map, phenoframe, parent = "N2xCB4856", tsize = 20)
}
\arguments{
\item{cross}{The cross object used for the mapping (output of mergepheno() function)}

\item{map}{The mapping output for a single trait}

\item{phenoframe}{The dataframe used to make the cross object}

\item{parent}{The parental cross, either "N2xCB4856" (default), "N2xLSJ2",
or "AF16xHK104"}

\item{tsize}{The theme size of the text for the plot, default is 20}
}
\value{
A boxplot of the phenotype by genotype split at each peak marker in a
mapping, including parental phenotypes
}
\description{
Plot the phenotype by genotype split at each peak marker in a mapping
next to parental phenotype split
}
