% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{annotate_lods}
\alias{annotate_lods}
\title{Annotate LOD peaks with variance explained, effect size, and confidence
interval bounds}
\usage{
annotate_lods(lods, cross)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{cross}{The cross object used for the original mapping}
}
\value{
The annotated lods data frame with information added for peak markers
of each iteration
}
\description{
Annotate LOD peaks with variance explained, effect size, and confidence
interval bounds
}

