% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{effectplot}
\alias{effectplot}
\title{Plot effect size by marker}
\usage{
effectplot(cross, map)
}
\arguments{
\item{cross}{The cross object used for the mapping}

\item{map}{The mapping output for a single trait}
}
\value{
A line plot of the effect size at each marker
}
\description{
Plot effect size by marker
}

