% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{get_peak_fdr}
\alias{get_peak_fdr}
\title{Get the FDR value for a particular mapping by phenotype permutation}
\usage{
get_peak_fdr(lods, cross, perms = 1000, doGPU = F)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{cross}{The original cross object used to perform the mapping}

\item{perms}{}

\item{doGPU}{Boolean, whether to use the gputools package to speed up,
mapping. This can only be set to \code{TRUE} on machines with an NVIDEA
graphics card with the gputools package installed. Defaults to \code{FALSE}.}
}
\value{
The value of the 5% FDR threshold
}
\description{
Get the FDR value for a particular mapping by phenotype permutation
}

