% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{get_peaks_above_thresh}
\alias{get_peaks_above_thresh}
\title{Return entries from a LOD data frame that are above the threshold value}
\usage{
get_peaks_above_thresh(lods, threshold)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{threshold}{The FDR threshold value used to determine significant peaks}
}
\value{
A subset of the \code{lods} data for only markers with a LOD score
above the threshold value
}
\description{
If more two or more markers from the same trait share the max LOD score, only
the first marker will be taken to be the peak.
}

