% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{get_pheno_resids}
\alias{get_pheno_resids}
\title{Regress genotype from phenotype and resturn the residual phenotype values}
\usage{
get_pheno_resids(lods, cross, threshold, intercept = FALSE)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{cross}{The cross object used for the original mapping}

\item{threshold}{The FDR threshold value used to determine significant peaks}

\item{intercept}{Boolean stating whether or not to include intercept term in
linear model. Defaults to \code{FALSE}.}
}
\value{
The residuals of the phenotypes
}
\description{
Regress genotype from phenotype and resturn the residual phenotype values
}

