% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{lodmatrix2scanone}
\alias{lodmatrix2scanone}
\title{Convert LODmatrix to scanone object}
\usage{
lodmatrix2scanone(lods, cross)
}
\arguments{
\item{lods}{A data frame output by the mapping functions to be converted to a
\code{scanone} object}

\item{cross}{An example cross object from which to extract scanone skeleton}
}
\value{
A scanone onject with the resultant mapping data in the lod column
}
\description{
Convert LODmatrix to scanone object
}

