% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postmapping.R
\name{maxpeaks}
\alias{maxpeaks}
\title{Get max lod score and SNP index for each mapped trait}
\usage{
maxpeaks(cross, lods)
}
\arguments{
\item{cross}{An example cross object from which to extract scanone skeleton}

\item{lods}{A data frame output by the mapping functions to be converted to a}
}
\value{
The genotype matrix, encoded as -1 or 1 for genotype
}
\description{
Get max lod score and SNP index for each mapped trait
}

