% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{pxgplot}
\alias{pxgplot}
\title{Plot the phenotype by genotype split at each peak marker in a mapping}
\usage{
pxgplot(cross, map)
}
\arguments{
\item{cross}{The cross object used for the mapping}

\item{map}{The mapping output for a single trait}
}
\value{
A boxplot of the phenotype by genotype split at each peak marker in a
mapping
}
\description{
Plot the phenotype by genotype split at each peak marker in a mapping
}

